Planner.CPP
As coded the find conflict option would always return that there was a conflict. So in the changed the planner.cpp file. I added code that would allow this function to run. Just replace the enclosed code with that currently in the Planner.cpp. It should compare the different times of all elements in the planner linked list returning whether there is a conflict or not.